+++ packages/gatsby-source-shopify/src/queries.js
@@ -34,7 +34,9 @@ export const ARTICLES_QUERY = `
           contentHtml
           excerpt
           excerptHtml
+          handle
           id
+          handle
           image {
             altText
             id
@@ -63,7 +65,9 @@ export const BLOGS_QUERY = `
       edges {
         cursor
         node {
+          handle
           id
+          handle
           title
           url
         }
@@ -227,23 +231,36 @@ export const PRODUCTS_QUERY = `
   }
 `
 
+export const SHOP_DETAILS_QUERY = `
+query GetShop {
+  shop {
+    description
+    moneyFormat
+    name
+  }
+}
+`
+
 export const SHOP_POLICIES_QUERY = `
   query GetPolicies {
     shop {
       privacyPolicy {
         body
+        handle
         id
         title
         url
       }
       refundPolicy {
         body
+        handle
         id
         title
         url
       }
       termsOfService {
         body
+        handle
         id
         title
         url