+++ packages/gatsby/src/query/index.js
@@ -8,6 +8,7 @@ const { boundActionCreators } = require(`../redux/actions`)
 const report = require(`gatsby-cli/lib/reporter`)
 const queryQueue = require(`./queue`)
 const { GraphQLRunner } = require(`./graphql-runner`)
+const pageDataUtil = require(`../utils/page-data`)
 
 const seenIdsWithoutDataDependencies = new Set()
 let queuedDirtyActions = []
@@ -27,7 +28,7 @@ emitter.on(`DELETE_NODE`, action => {
   queuedDirtyActions.push({ payload: action.payload })
 })
 
-/////////////////////////////////////////////////////////////////////
+// ///////////////////////////////////////////////////////////////////
 // Calculate dirty static/page queries
 
 const popExtractedQueries = () => {
@@ -97,6 +98,9 @@ const popNodeQueries = state => {
 
     return dirtyIds
   }, new Set())
+
+  boundActionCreators.deleteComponentsDependencies([...uniqDirties])
+
   queuedDirtyActions = []
   return uniqDirties
 }
@@ -162,8 +166,10 @@ const processQueries = async (
   queryJobs,
   { activity, graphqlRunner, graphqlTracing }
 ) => {
-  const queue = queryQueue.createBuildQueue(graphqlRunner, { graphqlTracing })
-  await queryQueue.processBatch(queue, queryJobs, activity)
+  const queue = queryQueue.createAppropriateQueue(graphqlRunner, {
+    graphqlTracing,
+  })
+  return queryQueue.processBatch(queue, queryJobs, activity)
 }
 
 const createStaticQueryJob = (state, queryId) => {
@@ -236,55 +242,6 @@ const processPageQueries = async (
   )
 }
 
-const getInitialQueryProcessors = ({
-  parentSpan,
-  graphqlRunner,
-  graphqlTracing,
-} = {}) => {
-  const state = store.getState()
-  const queryIds = calcInitialDirtyQueryIds(state)
-  const { staticQueryIds, pageQueryIds } = groupQueryIds(queryIds)
-
-  const queryjobsCount =
-    _.filter(pageQueryIds.map(id => state.pages.get(id))).length +
-    staticQueryIds.length
-
-  let activity = null
-  let processedQueuesCount = 0
-  const createProcessor = (fn, queryIds) => async () => {
-    if (!activity) {
-      activity = createQueryRunningActivity(queryjobsCount, parentSpan)
-    }
-
-    await fn(queryIds, { state, activity, graphqlRunner, graphqlTracing })
-
-    processedQueuesCount++
-    // if both page and static queries are done, finish activity
-    if (processedQueuesCount === 2) {
-      activity.done()
-    }
-  }
-
-  return {
-    processStaticQueries: createProcessor(processStaticQueries, staticQueryIds),
-    processPageQueries: createProcessor(processPageQueries, pageQueryIds),
-    pageQueryIds,
-  }
-}
-
-const initialProcessQueries = async ({ parentSpan, graphqlTracing } = {}) => {
-  const {
-    pageQueryIds,
-    processPageQueries,
-    processStaticQueries,
-  } = getInitialQueryProcessors({ parentSpan, graphqlTracing })
-
-  await processStaticQueries()
-  await processPageQueries()
-
-  return { pageQueryIds }
-}
-
 const createPageQueryJob = (state, page) => {
   const component = state.components.get(page.componentPath)
   const { path, componentPath, context } = page
@@ -301,7 +258,7 @@ const createPageQueryJob = (state, page) => {
   }
 }
 
-/////////////////////////////////////////////////////////////////////
+// ///////////////////////////////////////////////////////////////////
 // Listener for gatsby develop
 
 // Initialized via `startListening`
@@ -352,6 +309,7 @@ const startListeningToDevelopQueue = ({ graphqlTracing } = {}) => {
     const activity = createQueryRunningActivity(queryJobs.length)
 
     const onFinish = (...arg) => {
+      pageDataUtil.enqueueFlush()
       activity.done()
       return callback(...arg)
     }
@@ -408,11 +366,10 @@ const enqueueExtractedPageComponent = componentPath => {
 
 module.exports = {
   calcInitialDirtyQueryIds,
+  calcDirtyQueryIds,
   processPageQueries,
   processStaticQueries,
   groupQueryIds,
-  initialProcessQueries,
-  getInitialQueryProcessors,
   startListeningToDevelopQueue,
   runQueuedQueries,
   enqueueExtractedQueryId,