+++ packages/gatsby/src/state-machines/query-running/actions.ts
@@ -1,0 +1,25 @@
+import { IQueryRunningContext } from "./types"
+import { DoneInvokeEvent, assign, ActionFunctionMap } from "xstate"
+import { enqueueFlush } from "../../utils/page-data"
+
+export const flushPageData = (): void => {
+  enqueueFlush()
+}
+
+export const assignDirtyQueries = assign<
+  IQueryRunningContext,
+  DoneInvokeEvent<any>
+>((_context, { data }) => {
+  const { queryIds } = data
+  return {
+    queryIds,
+  }
+})
+
+export const queryActions: ActionFunctionMap<
+  IQueryRunningContext,
+  DoneInvokeEvent<any>
+> = {
+  assignDirtyQueries,
+  flushPageData,
+}