+++ packages/gatsby-admin/CHANGELOG.md
@@ -3,6 +3,218 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.100](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.99...gatsby-admin@0.1.100) (2020-07-24)
+
+### Bug Fixes
+
+- **gatsby-admin:** show error messages in the interface ([#25944](https://github.com/gatsbyjs/gatsby/issues/25944)) ([491197f](https://github.com/gatsbyjs/gatsby/commit/491197f))
+- **gatsby-admin:** small design tweaks ([#25971](https://github.com/gatsbyjs/gatsby/issues/25971)) ([e699564](https://github.com/gatsbyjs/gatsby/commit/e699564))
+
+## [0.1.99](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.98...gatsby-admin@0.1.99) (2020-07-23)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.98](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.97...gatsby-admin@0.1.98) (2020-07-22)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.97](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.96...gatsby-admin@0.1.97) (2020-07-21)
+
+### Features
+
+- **gatsby-admin:** list all pages of site ([#25744](https://github.com/gatsbyjs/gatsby/issues/25744)) ([a0c70d0](https://github.com/gatsbyjs/gatsby/commit/a0c70d0))
+- **gatsby-admin:** plugin search ([#25903](https://github.com/gatsbyjs/gatsby/issues/25903)) ([0506123](https://github.com/gatsbyjs/gatsby/commit/0506123))
+
+## [0.1.96](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.95...gatsby-admin@0.1.96) (2020-07-20)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.95](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.94...gatsby-admin@0.1.95) (2020-07-20)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.94](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.93...gatsby-admin@0.1.94) (2020-07-20)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.93](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.92...gatsby-admin@0.1.93) (2020-07-17)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.92](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.91...gatsby-admin@0.1.92) (2020-07-15)
+
+### Features
+
+- **gatsby-admin:** move to dropdown for plugin/theme actions ([#25598](https://github.com/gatsbyjs/gatsby/issues/25598)) ([9f84f11](https://github.com/gatsbyjs/gatsby/commit/9f84f11))
+
+## [0.1.91](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.90...gatsby-admin@0.1.91) (2020-07-10)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.90](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.89...gatsby-admin@0.1.90) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.89](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.88...gatsby-admin@0.1.89) (2020-07-09)
+
+### Features
+
+- **gatsby-admin:** remove all custom color treatment ([#25597](https://github.com/gatsbyjs/gatsby/issues/25597)) ([61b4291](https://github.com/gatsbyjs/gatsby/commit/61b4291))
+
+## [0.1.88](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.87...gatsby-admin@0.1.88) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.87](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.86...gatsby-admin@0.1.87) (2020-07-07)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.86](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.85...gatsby-admin@0.1.86) (2020-07-06)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.85](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.84...gatsby-admin@0.1.85) (2020-07-03)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.84](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.83...gatsby-admin@0.1.84) (2020-07-03)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.83](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.82...gatsby-admin@0.1.83) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.82](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.81...gatsby-admin@0.1.82) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.81](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.80...gatsby-admin@0.1.81) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.80](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.79...gatsby-admin@0.1.80) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.79](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.78...gatsby-admin@0.1.79) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.78](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.77...gatsby-admin@0.1.78) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.77](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.76...gatsby-admin@0.1.77) (2020-07-01)
+
+### Bug Fixes
+
+- update packages ([#25381](https://github.com/gatsbyjs/gatsby/issues/25381)) ([622e3c1](https://github.com/gatsbyjs/gatsby/commit/622e3c1))
+
+## [0.1.76](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.75...gatsby-admin@0.1.76) (2020-06-29)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.75](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.74...gatsby-admin@0.1.75) (2020-06-24)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.74](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.73...gatsby-admin@0.1.74) (2020-06-23)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.73](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.72...gatsby-admin@0.1.73) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.72](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.71...gatsby-admin@0.1.72) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.71](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.70...gatsby-admin@0.1.71) (2020-06-20)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.70](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.69...gatsby-admin@0.1.70) (2020-06-19)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.69](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.68...gatsby-admin@0.1.69) (2020-06-19)
+
+### Bug Fixes
+
+- **gatsby-admin:** add missing `nodemon` dependency ([#25014](https://github.com/gatsbyjs/gatsby/issues/25014)) ([68ca7da](https://github.com/gatsbyjs/gatsby/commit/68ca7da)), closes [#25007](https://github.com/gatsbyjs/gatsby/issues/25007)
+
+## [0.1.68](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.67...gatsby-admin@0.1.68) (2020-06-15)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.67](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.66...gatsby-admin@0.1.67) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.66](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.65...gatsby-admin@0.1.66) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.65](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.64...gatsby-admin@0.1.65) (2020-06-05)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.64](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.63...gatsby-admin@0.1.64) (2020-06-04)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.63](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.62...gatsby-admin@0.1.63) (2020-06-04)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.62](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.61...gatsby-admin@0.1.62) (2020-06-04)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.61](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.60...gatsby-admin@0.1.61) (2020-06-03)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.60](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.59...gatsby-admin@0.1.60) (2020-06-03)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.59](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.58...gatsby-admin@0.1.59) (2020-06-03)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.58](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.57...gatsby-admin@0.1.58) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.57](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.56...gatsby-admin@0.1.57) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.56](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.55...gatsby-admin@0.1.56) (2020-05-31)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.55](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.54...gatsby-admin@0.1.55) (2020-05-31)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.54](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.53...gatsby-admin@0.1.54) (2020-05-30)
+
+### Features
+
+- **gatsby:** serve Admin from develop parent process ([#23734](https://github.com/gatsbyjs/gatsby/issues/23734)) ([a6eb21b](https://github.com/gatsbyjs/gatsby/commit/a6eb21b))
+
+## [0.1.53](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.52...gatsby-admin@0.1.53) (2020-05-28)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.52](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.51...gatsby-admin@0.1.52) (2020-05-26)
+
+**Note:** Version bump only for package gatsby-admin
+
 ## [0.1.51](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.50...gatsby-admin@0.1.51) (2020-05-25)
 
 **Note:** Version bump only for package gatsby-admin